pyproject.toml
[project]
name = "desktop-archives-scraper"
version = "0.1.0"
description = "Windows desktop ingestion worker for UCSC archives backlog, sharing production DB schema with archives_scraper"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"beautifulsoup4>=4.13.0",
	"click>=8.1.0",
	"httpx>=0.28.0",
	"lxml>=5.3.0",
	"markdown>=3.7",
	"mammoth>=1.8.0",
	"numpy>=2.2.0",
	"ocrmypdf>=16.10.0",
	"openpyxl>=3.1.0",
	"pandas>=2.2.0",
	"pgvector>=0.4.0",
	"pillow>=11.0.0",
	"psycopg[binary]>=3.2.0",
	"pymupdf>=1.25.0",
	"python-docx>=1.1.0",
	"python-dotenv>=1.0.0",
	"python-pptx>=1.0.2",
	"pytesseract>=0.3.13",
	"sentence-transformers>=3.3.0",
	"sqlalchemy>=2.0.0",
	"striprtf>=0.0.29",
	"unidecode>=1.3.8",
	"xlrd>=2.0.1",
]

[project.scripts]
desktop-archives-scraper = "desktop_archives_scraper.cli:main"
